updat

+++ asm/src/main/java/com/example/asm/dao/CategoryDAO.java
@@ -3,5 +3,5 @@ package com.example.asm.dao;
 import com.example.asm.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryDAO extends JpaRepository<Category, Long> {
+public interface CategoryDAO extends JpaRepository<Category, Integer> {
 }
